Rewrite appendBytes() using a PHP match
Nicer, and more consistent with some other code in the encode() method.

Invalid modes are really not expected here, as this private method is called only by encode(),
and the mode is determined right at the beginning of encode() by calling chooseMode().

Though, if an invalid mode ever comes here, the match would throw an UnhandledMatchError.